+++ Modules/Ruleset.Utility/Public/Invoke-Process.ps1
@@ -35,12 +35,8 @@ Run process with or without arguments, set wait time and capture output.
 If the target process results in an error, error message is formatted and shown in addition
 to standard output if any.
 
-.PARAMETER Application
-Executable name.
-Should be in PATH environment variable to resolve.
-
 .PARAMETER Path
-The path to application or document to start.
+Executable name or path to application to which to start.
 Wildcard characters and relative paths are supported.
 
 .PARAMETER ArgumentList
@@ -50,37 +46,36 @@ A collection of command-line arguments to use when starting the application
 Whether to use the operating system shell to start the process
 
 .PARAMETER Wait
-Number of milliseconds to wait for the associated process to exit
-Default is 0, which means wait indefinitely
+The amount of time, in milliseconds, to wait for the associated process to exit.
+Value 0 means an immediate return, and a value of -1 specifies an infinite wait.
+The default wait time is 10 seconds.
 
 .PARAMETER Format
-If specified, formats standard output into INFO messages
+If specified, each line of application output is redirected to new information stream
 
 .EXAMPLE
-PS> Get-ProcessOutput -Path "git.exe" -ArgumentList "status" -NoNewWindow -Wait 3000
+PS> Invoke-Process git.exe -ArgumentList "status" -NoNewWindow -Wait 3000
 
 .INPUTS
-None. You cannot pipe objects to Get-ProcessOutput
+None. You cannot pipe objects to Invoke-Process
 
 .OUTPUTS
 [string] If the "Format" parameter is not specified
 
 .NOTES
-TODO: Function needs improvements and more test cases
-TODO: consider renaming to Format-ProcessOutput
+TODO: Because of uncertain output this function needs a lot of improvements and a lot more test cases
+to handle variable varieties of process outputs.
 #>
-function Get-ProcessOutput
+function Invoke-Process
 {
-	[CmdletBinding(PositionalBinding = $false, DefaultParameterSetName = "Program",
-		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Utility/Help/en-US/Get-ProcessOutput.md")]
+	[CmdletBinding(PositionalBinding = $false,
+		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Utility/Help/en-US/Invoke-Process.md")]
 	[OutputType([string])]
 	Param (
-		[Parameter(Mandatory = $true, Position = 0, ParameterSetName = "Program")]
-		[string] $Application,
-
-		[Parameter(Mandatory = $true, ParameterSetName = "Path")]
+		[Parameter(Mandatory = $true, Position = 0)]
 		[SupportsWildcards()]
-		[System.IO.FileInfo] $Path,
+		[Alias("Application", "FilePath")]
+		[string] $Path,
 
 		[Parameter()]
 		[string] $ArgumentList,
@@ -89,42 +84,54 @@ function Get-ProcessOutput
 		[switch] $NoNewWindow,
 
 		[Parameter()]
-		[uint32] $Wait = 0,
+		[int32] $Wait = 10000,
 
 		[Parameter()]
 		[switch] $Format
 	)
 
-	if ($Path)
+	$CommandName = Split-Path -Path $Path -Leaf
+
+	# TODO: May return unexpected application if multiple matches exist
+	# [System.Management.Automation.ApplicationInfo]
+	$Command = Get-Command -Name $CommandName -CommandType Application -ErrorAction Ignore
+
+	if (!$Command)
 	{
 		[System.IO.FileInfo] $FilePath = Resolve-FileSystem $Path -File
-		if (!($FilePath -and $FilePath.Exists))
+
+		if ($FilePath -and $FilePath.Exists)
 		{
+			$Command = Get-Command -Name $FilePath.FullName -CommandType Application -ErrorAction Ignore
+
+			if (!$Command)
+			{
+				Write-Error -Category ObjectNotFound -TargetObject $FilePath.FullName -Message "Application was not found: $($FilePath.FullName)"
+				return
+			}
+		}
+		else
+		{
+			Write-Error -Category ObjectNotFound -TargetObject $CommandName -Message "Application '$CommandName' was not found"
 			return
 		}
 	}
-	else
-	{
-		[System.Management.Automation.ApplicationInfo] $Command = Get-Command -Name $Application -CommandType Application -ErrorAction SilentlyContinue
 
-		if (!$Command)
-		{
-			Write-Error -Category ObjectNotFound -TargetObject $Application -Message "The application '$Application' was not found"
-			return
-		}
-
-		[System.IO.FileInfo] $FilePath = $Command.Source
-	}
-
+	$CommandName = $Command.Name
 	$Process = New-Object System.Diagnostics.Process
+
 	# The application or document to start
-	$Process.StartInfo.FileName = $FilePath
+	$Process.StartInfo.FileName = $Command.Path
+
 	# Whether the textual output of an application is written to the StandardOutput stream
 	$Process.StartInfo.RedirectStandardOutput = $true
+
 	# Whether the error output of an application is written to the StandardError stream
 	$Process.StartInfo.RedirectStandardError = $true
+
 	# Whether to start the process in a new window
 	$Process.StartInfo.CreateNoWindow = $NoNewWindow
+
 	# Whether to use the operating system shell to start the process
 	$Process.StartInfo.UseShellExecute = $false
 
@@ -150,30 +157,40 @@ function Get-ProcessOutput
 	# 	-Action $ScripBlock -EventName "OutputDataReceived" `
 	# 	-MessageData $StdOutBuilder
 
-	if (!$Process.Start())
+	try
 	{
-		Write-Warning -Message "Start process '$($Process.ProcessName)' failed"
-		return
+		Write-Verbose -Message "Starting process '$CommandName'"
+
+		# true if a process resource is started; false if no new process resource is started
+		if (!$Process.Start())
+		{
+			Write-Error -Category InvalidResult -TargetObject $Process -Message "Starting process '$CommandName' failed"
+			return
+		}
+	}
+	catch
+	{
+		Write-Error -Category $_.CategoryInfo.Category -TargetObject $_.TargetObject -Message $_.Exception.Message
 	}
 
 	# NOTE: Part of commented code above
 	# $Process.BeginOutputReadLine()
 
-	if ($Wait -gt 0)
+	if ($Wait -ge 0)
 	{
-		# true if the associated process has exited; otherwise, false
-		$Status = $Process.WaitForExit([int32] $Wait)
+		Write-Information -Tags "User" -MessageData "Waiting up to $($Wait / 1000) seconds for Process '$CommandName' to finish ..."
 
-		if (!$Status)
+		# true if the associated process has exited; otherwise, false
+		if (!$Process.WaitForExit($Wait))
 		{
-			Write-Warning -Message "Process '$($Process.ProcessName)' is taking too long, aborting..."
+			Write-Warning -Message "Process '$CommandName' is taking too long, aborting ..."
 			$Process.Kill()
 			return
 		}
 	}
 	else
 	{
-		Write-Debug -Message "[$($MyInvocation.InvocationName)] Waiting for Process '$($Process.ProcessName)' to finish..."
+		Write-Information -Tags "User" -MessageData "Waiting infinitely for Process '$CommandName' to finish ..."
 
 		$Process.WaitForExit()
 
@@ -186,10 +203,15 @@ function Get-ProcessOutput
 
 	if ($Format)
 	{
-		# Reads a line of characters from the current stream and returns the data as a string.
-		while ($null -ne ($StreamLine = $Process.StandardOutput.ReadLine()))
+		Write-Verbose -Message "Formatting process output"
+
+		# true if the current stream position is at the end of the stream
+		while (!$Process.StandardOutput.EndOfStream)
 		{
-			if ($StreamLine)
+			# Reads a line of characters from the current stream and returns the data as [string]
+			$StreamLine = $Process.StandardOutput.ReadLine()
+
+			if (![string]::IsNullOrEmpty($StreamLine))
 			{
 				Write-Information -Tags "User" -MessageData "INFO: $StreamLine"
 			}
@@ -197,14 +219,15 @@ function Get-ProcessOutput
 	}
 	else
 	{
+		Write-Verbose -Message "Getting process output"
 		$Process.StandardOutput.ReadToEnd()
 	}
 
 	# Reads all characters from the current position to the end of the stream (returns [string])
 	$StandardError = $Process.StandardError.ReadToEnd()
 
-	if ($StandardError)
+	if (![string]::IsNullOrEmpty($StandardError))
 	{
-		Write-Error -Category FromStdErr -Message $StandardError -TargetObject $Process
+		Write-Error -Category FromStdErr -TargetObject $Process -Message $StandardError
 	}
 }